refactor: extract TransferProcessors
