+++ core/control-plane/control-plane-transfer-manager/src/main/java/org/eclipse/edc/connector/controlplane/transfer/TransferManagerExtension.java
@@ -14,23 +14,20 @@
 
 package org.eclipse.edc.connector.controlplane.transfer;
 
-import org.eclipse.edc.connector.controlplane.asset.spi.index.DataAddressResolver;
 import org.eclipse.edc.connector.controlplane.policy.spi.store.PolicyArchive;
 import org.eclipse.edc.connector.controlplane.transfer.process.TransferProcessManagerImpl;
 import org.eclipse.edc.connector.controlplane.transfer.spi.TransferProcessManager;
 import org.eclipse.edc.connector.controlplane.transfer.spi.TransferProcessPendingGuard;
-import org.eclipse.edc.connector.controlplane.transfer.spi.flow.DataFlowController;
+import org.eclipse.edc.connector.controlplane.transfer.spi.TransferProcessors;
 import org.eclipse.edc.connector.controlplane.transfer.spi.observe.TransferProcessObservable;
 import org.eclipse.edc.connector.controlplane.transfer.spi.retry.TransferWaitStrategy;
 import org.eclipse.edc.connector.controlplane.transfer.spi.store.TransferProcessStore;
 import org.eclipse.edc.connector.controlplane.transfer.spi.types.DataAddressStore;
-import org.eclipse.edc.protocol.spi.DataspaceProfileContextRegistry;
 import org.eclipse.edc.runtime.metamodel.annotation.Configuration;
 import org.eclipse.edc.runtime.metamodel.annotation.Extension;
 import org.eclipse.edc.runtime.metamodel.annotation.Inject;
 import org.eclipse.edc.runtime.metamodel.annotation.Provides;
 import org.eclipse.edc.runtime.metamodel.annotation.SettingContext;
-import org.eclipse.edc.spi.message.RemoteMessageDispatcherRegistry;
 import org.eclipse.edc.spi.system.ExecutorInstrumentation;
 import org.eclipse.edc.spi.system.ServiceExtension;
 import org.eclipse.edc.spi.system.ServiceExtensionContext;
@@ -55,27 +52,21 @@ public class TransferManagerExtension implements ServiceExtension {
     @Inject
     private TransferProcessStore transferProcessStore;
     @Inject
-    private DataFlowController dataFlowController;
-    @Inject
     private TransferProcessObservable observable;
     @Inject
     private PolicyArchive policyArchive;
     @Inject
-    private RemoteMessageDispatcherRegistry dispatcherRegistry;
-    @Inject
-    private DataAddressResolver addressResolver;
-    @Inject
     private Clock clock;
     @Inject
     private Telemetry telemetry;
-    @Inject
-    private DataspaceProfileContextRegistry dataspaceProfileContextRegistry;
     @Inject
     private TransferProcessPendingGuard pendingGuard;
     @Inject
     private ExecutorInstrumentation executorInstrumentation;
     @Inject
     private DataAddressStore dataAddressStore;
+    @Inject
+    private TransferProcessors transferProcessors;
 
     private TransferProcessManagerImpl processManager;
 
@@ -94,8 +85,6 @@ public class TransferManagerExtension implements ServiceExtension {
 
         processManager = TransferProcessManagerImpl.Builder.newInstance()
                 .waitStrategy(waitStrategy)
-                .dataFlowController(dataFlowController)
-                .dispatcherRegistry(dispatcherRegistry)
                 .monitor(monitor)
                 .telemetry(telemetry)
                 .executorInstrumentation(executorInstrumentation)
@@ -104,11 +93,10 @@ public class TransferManagerExtension implements ServiceExtension {
                 .store(transferProcessStore)
                 .policyArchive(policyArchive)
                 .batchSize(stateMachineConfiguration.batchSize())
-                .addressResolver(addressResolver)
                 .entityRetryProcessConfiguration(entityRetryProcessConfiguration)
-                .dataspaceProfileContextRegistry(dataspaceProfileContextRegistry)
                 .pendingGuard(pendingGuard)
                 .dataAddressStore(dataAddressStore)
+                .transferProcessors(transferProcessors)
                 .build();
 
         context.registerService(TransferProcessManager.class, processManager);
