+++ core/control-plane/control-plane-transfer/src/main/java/org/eclipse/edc/connector/controlplane/transfer/TransferCoreExtension.java
@@ -16,19 +16,35 @@
 package org.eclipse.edc.connector.controlplane.transfer;
 
 import jakarta.json.Json;
+import org.eclipse.edc.connector.controlplane.asset.spi.index.DataAddressResolver;
+import org.eclipse.edc.connector.controlplane.policy.spi.store.PolicyArchive;
 import org.eclipse.edc.connector.controlplane.transfer.edr.DataAddressToEndpointDataReferenceTransformer;
 import org.eclipse.edc.connector.controlplane.transfer.listener.TransferProcessEventListener;
+import org.eclipse.edc.connector.controlplane.transfer.processors.TransferProcessorsImpl;
+import org.eclipse.edc.connector.controlplane.transfer.spi.TransferProcessors;
+import org.eclipse.edc.connector.controlplane.transfer.spi.flow.DataFlowController;
 import org.eclipse.edc.connector.controlplane.transfer.spi.observe.TransferProcessObservable;
+import org.eclipse.edc.connector.controlplane.transfer.spi.store.TransferProcessStore;
+import org.eclipse.edc.connector.controlplane.transfer.spi.types.DataAddressStore;
+import org.eclipse.edc.protocol.spi.DataspaceProfileContextRegistry;
+import org.eclipse.edc.runtime.metamodel.annotation.Configuration;
 import org.eclipse.edc.runtime.metamodel.annotation.Extension;
 import org.eclipse.edc.runtime.metamodel.annotation.Inject;
+import org.eclipse.edc.runtime.metamodel.annotation.Provider;
+import org.eclipse.edc.runtime.metamodel.annotation.SettingContext;
 import org.eclipse.edc.spi.event.EventRouter;
+import org.eclipse.edc.spi.message.RemoteMessageDispatcherRegistry;
+import org.eclipse.edc.spi.monitor.Monitor;
 import org.eclipse.edc.spi.system.ServiceExtension;
 import org.eclipse.edc.spi.system.ServiceExtensionContext;
 import org.eclipse.edc.spi.types.TypeManager;
+import org.eclipse.edc.statemachine.StateMachineConfiguration;
+import org.eclipse.edc.statemachine.retry.EntityRetryProcessFactory;
 import org.eclipse.edc.transform.spi.TypeTransformerRegistry;
 import org.eclipse.edc.transform.transformer.edc.from.JsonObjectFromDataAddressTransformer;
 import org.eclipse.edc.transform.transformer.edc.to.JsonObjectToDataAddressTransformer;
 
+import java.time.Clock;
 import java.util.Collections;
 
 import static org.eclipse.edc.spi.constants.CoreConstants.JSON_LD;
@@ -41,6 +57,10 @@ public class TransferCoreExtension implements ServiceExtension {
 
     public static final String NAME = "Transfer Core";
 
+    @SettingContext("edc.transfer")
+    @Configuration
+    private StateMachineConfiguration stateMachineConfiguration;
+
     @Inject
     private TransferProcessObservable observable;
     @Inject
@@ -49,6 +69,24 @@ public class TransferCoreExtension implements ServiceExtension {
     private TypeTransformerRegistry typeTransformerRegistry;
     @Inject
     private TypeManager typeManager;
+    @Inject
+    private PolicyArchive policyArchive;
+    @Inject
+    private DataFlowController dataFlowController;
+    @Inject
+    private DataAddressStore dataAddressStore;
+    @Inject
+    private TransferProcessStore store;
+    @Inject
+    private Monitor monitor;
+    @Inject
+    private DataAddressResolver addressResolver;
+    @Inject
+    private DataspaceProfileContextRegistry dataspaceProfileContextRegistry;
+    @Inject
+    private RemoteMessageDispatcherRegistry dispatcherRegistry;
+    @Inject
+    private Clock clock;
 
     @Override
     public String name() {
@@ -63,6 +101,14 @@ public class TransferCoreExtension implements ServiceExtension {
         typeTransformerRegistry.register(new JsonObjectFromDataAddressTransformer(builderFactory, typeManager, JSON_LD));
 
         observable.registerListener(new TransferProcessEventListener(eventRouter));
+    }
+
+    @Provider
+    public TransferProcessors transferProcessors() {
+        var entityRetryProcessConfiguration = stateMachineConfiguration.entityRetryProcessConfiguration();
+        var entityRetryProcessFactory = new EntityRetryProcessFactory(monitor, clock, entityRetryProcessConfiguration);
+        return new TransferProcessorsImpl(policyArchive, entityRetryProcessFactory, dataFlowController, dataAddressStore,
+                observable, store, monitor, addressResolver, dataspaceProfileContextRegistry, dispatcherRegistry);
     }
 
 }
