+++ extensions/common/iam/decentralized-claims/decentralized-claims-core/src/test/java/org/eclipse/edc/iam/decentralizedclaims/core/DidConfigProviderTest.java
@@ -42,6 +42,7 @@ class DidConfigProviderTest {
         var did = provider.apply("participantContextId");
 
         assertThat(did).isNull();
+        verify(monitor).severe(anyString());
     }
 
     @Test
